Separated creation of monsters and shopkeeper into a new class - ObjectCreator
