+++ src/controller/GameLogic.java
@@ -178,6 +178,7 @@ public class GameLogic {
                     window.getAnswerPanel().setVisible(false);
                 }
             }
+            /*
             else {
                 if (levelCreator.getLevel(counter.getLevel()).getEnemy().isBoss()) {
                     player.wrong(2);
@@ -195,6 +196,7 @@ public class GameLogic {
                     healthBar.updateHealth();
                 }
             }
+             */
         }
     }
