+++ src/view/SceneCreator.java
@@ -20,8 +20,6 @@ import java.util.LinkedList;
 public class SceneCreator {
     private LinkedList<JPanel> bgPanels;
     private LinkedList<JLabel> bgImages;
-    private LinkedList<JLabel> monsters;
-    private LinkedList<JLabel> shop;
     private LinkedList<JLabel> doors;
 
     private MainFrame mainFrame;
@@ -42,8 +40,6 @@ public class SceneCreator {
 
         bgPanels = new LinkedList<>();
         bgImages = new LinkedList<>();
-        monsters = new LinkedList<>();
-        shop = new LinkedList<>();
         doors = new LinkedList<>();
     }
 
@@ -53,8 +49,6 @@ public class SceneCreator {
     public void generateScenes() {
         createBackgrounds();
         createDoorObjects();
-        createMonsterObjects();
-        createShopKeeper();
     }
 
     /**
@@ -235,214 +229,6 @@ public class SceneCreator {
     }
 
     /**
-     * Creates a popup menu for all monsters and returns it for use
-     * @return the popup menu
-     */
-    public JPopupMenu createMonsterMenu() {
-        JPopupMenu monsterMenu = new JPopupMenu();
-
-        JMenuItem[] menuItems = new JMenuItem[3];
-        menuItems[0] = new JMenuItem("Look");
-        menuItems[0].addActionListener(actionHandler);
-        menuItems[0].setActionCommand("lookAtEnemy");
-
-        menuItems[1] = new JMenuItem("Talk");
-        menuItems[1].addActionListener(actionHandler);
-        menuItems[1].setActionCommand("talkToEnemy");
-
-        menuItems[2] = new JMenuItem("Attack");
-        menuItems[2].addActionListener(actionHandler);
-        menuItems[2].setActionCommand("attackEnemy");
-
-        for (JMenuItem mi : menuItems) {
-            monsterMenu.add(mi);
-        }
-
-        return monsterMenu;
-    }
-
-    /**
-     * Creates all monster objects (label)
-     */
-    public void createMonsterObjects() {
-        try {
-            BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream("resources/monsters/monsterImageLocation.txt")));
-            String str = br.readLine();
-            JPopupMenu popupMenu = createMonsterMenu();
-
-            int lvlNbr = 0;
-            while (str != null) {
-                JLabel lblMonster = new JLabel();
-
-                ImageIcon monsterIcon = resize(str, 150, 150);
-                lblMonster.setIcon(monsterIcon);
-
-                lblMonster.addMouseListener(new MouseListener() {
-                    @Override
-                    public void mouseClicked(MouseEvent e) {
-
-                    }
-
-                    @Override
-                    public void mousePressed(MouseEvent e) {
-                        if (SwingUtilities.isRightMouseButton(e)) {
-                            if (controller.getOutOfCombat()) {
-                                popupMenu.show(lblMonster, e.getX(), e.getY());
-                            }
-                        }
-                    }
-
-                    @Override
-                    public void mouseReleased(MouseEvent e) {
-
-                    }
-
-                    @Override
-                    public void mouseEntered(MouseEvent e) {
-
-                    }
-
-                    @Override
-                    public void mouseExited(MouseEvent e) {
-
-                    }
-                });
-
-                monsters.add(lvlNbr, lblMonster);
-                bgPanels.get(lvlNbr+2).add(monsters.get(lvlNbr));
-                mainFrame.populateAnswerPanel();
-                bgPanels.get(lvlNbr+2).add(bgImages.get(lvlNbr+2));
-
-                lvlNbr++;
-                str = br.readLine();
-            }
-            setMonsterBounds();
-            br.close();
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * This method creates the shop keeper (Label).
-     */
-    public void createShopKeeper() {
-        try {
-            BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream("resources/npc/shopkeeper/shopKeeperImageLocation.txt")));
-            String str = br.readLine();
-            JPopupMenu popupMenu = createShopMenu();
-
-            int lvlNbr = 0;
-            while (str != null) {
-                JLabel lblShopKeeper = new JLabel();
-
-                ImageIcon shopKeeperIcon = resize(str, 200, 350);
-                lblShopKeeper.setIcon(shopKeeperIcon);
-
-                lblShopKeeper.addMouseListener(new MouseListener() {
-                    @Override
-                    public void mouseClicked(MouseEvent e) {
-
-                    }
-
-                    @Override
-                    public void mousePressed(MouseEvent e) {
-                        if (SwingUtilities.isRightMouseButton(e)) {
-                            if (controller.getOutOfCombat()) {
-                                popupMenu.show(lblShopKeeper, e.getX(), e.getY());
-                            }
-                        }
-                    }
-
-                    @Override
-                    public void mouseReleased(MouseEvent e) {
-
-                    }
-
-                    @Override
-                    public void mouseEntered(MouseEvent e) {
-
-                    }
-
-                    @Override
-                    public void mouseExited(MouseEvent e) {
-
-                    }
-                });
-
-                shop.add(lvlNbr, lblShopKeeper);
-                bgPanels.get(22).add(shop.get(lvlNbr));
-                mainFrame.populateShopPanel();
-                bgPanels.get(22).add(bgImages.get(22));
-
-                lvlNbr++;
-                str = br.readLine();
-            }
-            setShopBounds();
-            br.close();
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Sets the shopkeepers bounds by reading a txt file.
-     */
-    public void setShopBounds() {
-        try {
-            BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream("resources/npc/shopkeeper/shopBounds.txt")));
-            String str = br.readLine();
-
-            while (str != null) {
-                int level = Integer.parseInt(str)-1; //LinkedList starts at 0
-                str = br.readLine();
-                int x = Integer.parseInt(str);
-                str = br.readLine();
-                int y = Integer.parseInt(str);
-                str = br.readLine();
-                int width = Integer.parseInt(str);
-                str = br.readLine();
-                int height = Integer.parseInt(str);
-                str = br.readLine();
-
-                shop.get(level).setBounds(x, y, width, height);
-            }
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    /**
-     * Creates a popup menu for the shop keeper and returns it for use
-     * @return the popup menu
-     */
-    public JPopupMenu createShopMenu() {
-        JPopupMenu shopMenu = new JPopupMenu();
-
-        JMenuItem[] menuItems = new JMenuItem[3];
-        menuItems[0] = new JMenuItem("Look");
-        menuItems[0].addActionListener(actionHandler);
-        menuItems[0].setActionCommand("lookAtShopKeeper");
-
-        menuItems[1] = new JMenuItem("Talk");
-        menuItems[1].addActionListener(actionHandler);
-        menuItems[1].setActionCommand("talkToShopKeeper");
-
-        menuItems[2] = new JMenuItem("Buy");
-        menuItems[2].addActionListener(actionHandler);
-        menuItems[2].setActionCommand("buyFromShopKeeper");
-
-        for (JMenuItem mi : menuItems) {
-            shopMenu.add(mi);
-        }
-
-        return shopMenu;
-    }
-
-    /**
      * Populates the answerPanel with questions depending on scene number
      * @param sceneNbr given scene number
      */
@@ -463,34 +249,6 @@ public class SceneCreator {
             default:
                 bgPanels.get(sceneNbr).add(bgImages.get(sceneNbr));
                 break;
-        }
-    }
-
-    /**
-     * Sets the bounds of the monsters by reading a file (includes dimensional values)
-     */
-    public void setMonsterBounds() {
-        try {
-            BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream("resources/monsters/monsterBounds.txt")));
-            String str = br.readLine();
-
-            while (str != null) {
-                int level = Integer.parseInt(str) - 1; //LinkedList starts at 0
-                str = br.readLine();
-                int x = Integer.parseInt(str);
-                str = br.readLine();
-                int y = Integer.parseInt(str);
-                str = br.readLine();
-                int width = Integer.parseInt(str);
-                str = br.readLine();
-                int height = Integer.parseInt(str);
-                str = br.readLine();
-
-                monsters.get(level).setBounds(x, y, width, height);
-            }
-        }
-        catch (IOException e) {
-            e.printStackTrace();
         }
     }
 
@@ -524,5 +282,13 @@ public class SceneCreator {
      */
     public JButton getBtnArrow() {
         return btnArrow;
+    }
+
+    public LinkedList<JPanel> getBgPanels() {
+        return bgPanels;
+    }
+
+    public LinkedList<JLabel> getBgImages() {
+        return bgImages;
     }
 }
